# nestjs-project/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/videos": {
      "post": {
        "description": "Pre-registers the video as a draft and opens a multipart upload in the storage. The client then requests presigned part URLs and uploads the file straight to the storage.",
        "operationId": "VideosController_startUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft created and multipart upload opened",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartUploadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Declared file size exceeds 10 GiB (VIDEO_TOO_LARGE)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Start a video upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{id}": {
      "get": {
        "description": "Returns the owner's view of a video: status, duration, metadata, slug and a presigned thumbnail URL. Clients poll it to follow processing → ready.",
        "operationId": "VideosController_getVideo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The video",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found or owned by another user (VIDEO_NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a video",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{id}/stream": {
      "get": {
        "description": "Returns a presigned GET URL for the original file, usable directly in <video src>. The storage serves Range requests (206). Only for ready videos.",
        "operationId": "VideosController_getStreamUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Streaming URL (valid for 1 hour)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoUrlResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found or owned by another user (VIDEO_NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not ready (VIDEO_NOT_READY)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a streaming URL",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{id}/download": {
      "get": {
        "description": "Returns a presigned GET URL that forces a download (Content-Disposition: attachment) with the original file name. Only for ready videos.",
        "operationId": "VideosController_getDownloadUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download URL (valid for 15 minutes)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoUrlResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found or owned by another user (VIDEO_NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not ready (VIDEO_NOT_READY)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a download URL",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{id}/upload/part-urls": {
      "post": {
        "description": "Returns presigned UploadPart URLs for a batch of part numbers. The client PUTs each part straight to the storage and keeps the ETag response header.",
        "operationId": "VideosController_getPartUrls",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartUrlsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Presigned URLs (valid for 1 hour)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartUrlsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (VALIDATION_ERROR) or part number above partCount (INVALID_PART_NUMBER)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found or owned by another user (VIDEO_NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not a draft (INVALID_VIDEO_STATUS)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get presigned part URLs",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{id}/upload/parts": {
      "get": {
        "description": "Lists the parts the storage already received, so the client can resume an interrupted upload. An expired upload marks the video as failed.",
        "operationId": "VideosController_listUploadedParts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Parts already received",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadedPartsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found or owned by another user (VIDEO_NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not a draft (INVALID_VIDEO_STATUS)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "The storage no longer has the multipart upload (UPLOAD_EXPIRED); the video becomes failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List uploaded parts",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{id}/upload/complete": {
      "post": {
        "description": "Completes the multipart upload, checks the final object (at most 10 GiB) and queues the video processing. The video becomes processing.",
        "operationId": "VideosController_completeUpload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Upload completed and processing queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteUploadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (VALIDATION_ERROR) or the storage rejected the part list (INVALID_UPLOAD_PARTS); the video stays a draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found or owned by another user (VIDEO_NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not a draft (INVALID_VIDEO_STATUS)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "The storage no longer has the multipart upload (UPLOAD_EXPIRED); the video becomes failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "The final object exceeds 10 GiB (VIDEO_TOO_LARGE); it is removed and the video becomes failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Complete the upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{id}/upload": {
      "delete": {
        "description": "Aborts the multipart upload in the storage and marks the video as failed (upload_aborted).",
        "operationId": "VideosController_abortUpload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Upload aborted"
          },
          "400": {
            "description": "Invalid id (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found or owned by another user (VIDEO_NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not a draft (INVALID_VIDEO_STATUS)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Abort the upload",
        "tags": [
          "videos"
        ]
      }
    }
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "StartUploadDto": {
        "type": "object",
        "properties": {}
      },
      "StartUploadResponseDto": {
        "type": "object",
        "properties": {
          "videoId": {
            "type": "string",
            "format": "uuid"
          },
          "uploadId": {
            "type": "string",
            "description": "Multipart upload id in the storage"
          },
          "slug": {
            "type": "string",
            "example": "dQw4w9WgXcQ",
            "minLength": 11,
            "maxLength": 11
          },
          "title": {
            "type": "string",
            "example": "Minha aula"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft"
            ],
            "example": "draft"
          },
          "partSize": {
            "type": "number",
            "example": 67108864,
            "description": "Part size in bytes"
          },
          "partCount": {
            "type": "number",
            "example": 3,
            "description": "Number of parts to upload"
          }
        },
        "required": [
          "videoId",
          "uploadId",
          "slug",
          "title",
          "status",
          "partSize",
          "partCount"
        ]
      },
      "VideoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "dQw4w9WgXcQ",
            "minLength": 11,
            "maxLength": 11
          },
          "title": {
            "type": "string",
            "example": "Minha aula"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "processing",
              "ready",
              "failed"
            ],
            "example": "ready"
          },
          "durationSeconds": {
            "type": "number",
            "nullable": true,
            "example": 3
          },
          "metadata": {
            "type": "object",
            "nullable": true,
            "description": "Curated ffprobe summary with `format`, `video` and `audio` (null without audio)",
            "additionalProperties": true
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true,
            "description": "Presigned GET URL of the JPEG thumbnail (valid for 1 hour)"
          },
          "processingError": {
            "type": "string",
            "nullable": true,
            "example": null,
            "description": "Failure reason when status is failed (e.g. upload_expired, invalid_video)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "status",
          "durationSeconds",
          "metadata",
          "thumbnailUrl",
          "processingError",
          "createdAt",
          "updatedAt"
        ]
      },
      "VideoUrlResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "Presigned GET URL served by the storage"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the URL expires"
          }
        },
        "required": [
          "url",
          "expiresAt"
        ]
      },
      "PartUrlsDto": {
        "type": "object",
        "properties": {}
      },
...
      "PartUrlsResponseDto": {
        "type": "object",
        "properties": {
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PartUrlDto"
            }
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the URLs expire"
          }
        },
        "required": [
          "parts",
          "expiresAt"
        ]
      },
...
      "CompleteUploadDto": {
        "type": "object",
        "properties": {}
      },
...
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      }
    }
  }
}